requirements.txt
autopep8
python-dotenv
git+https://github.com/xp-network/smartpy.git